# agent-web/src/test/resources/contracts/feature005-openapi.yaml
openapi: 3.1.0
info:
  title: HiAgent Feature005 Run and Citation API
  version: 0.5.0-design
  description: Design only. Extends the verified Feature004 baseline. Session/memory operations retain feature004-openapi.yaml.
    No new public corpus import API.
servers:
- url: /
security:
- BearerAuth: []
paths:
  /api/agent/tasks:
    post:
      operationId: createAgentRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '202':
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunAccepted'
          description: Response 202; application invariants are defined in the Feature005 contracts.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response 409; application invariants are defined in the Feature005 contracts.
        '429':
          $ref: '#/components/responses/Capacity'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Create an owned run; requireEvidence defaults to false
  /api/agent/tasks/{taskId}:
    parameters:
    - name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      operationId: getAgentRun
      responses:
        '200':
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunSnapshot'
          description: Response 200; application invariants are defined in the Feature005 contracts.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/Capacity'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Read committed run and citation snapshots
  /api/agent/tasks/{taskId}/steps:
    parameters:
    - name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      operationId: getAgentRunSteps
      responses:
        '200':
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                type: array
                maxItems: 256
                items:
                  $ref: '#/components/schemas/StepResponse'
          description: Response 200; application invariants are defined in the Feature005 contracts.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/Capacity'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Read bounded content-free retrieval diagnostics
  /api/agent/tasks/{taskId}/cancel:
    parameters:
    - name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    post:
      operationId: cancelAgentRun
      responses:
        '200':
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunSnapshot'
          description: Response 200; application invariants are defined in the Feature005 contracts.
        '202':
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunSnapshot'
          description: Response 202; application invariants are defined in the Feature005 contracts.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/Capacity'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Request cooperative cancellation; return current snapshot
  /api/agent/tasks/{taskId}/events:
    parameters:
    - name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      operationId: subscribeAgentRunEvents
      parameters:
      - name: Last-Event-ID
        in: header
        required: false
        schema:
          type: string
          pattern: ^(0|[1-9][0-9]{0,18})$
          maxLength: 19
      responses:
        '200':
          headers:
            Cache-Control:
              schema:
                const: no-store
          content:
            text/event-stream:
              schema:
                type: string
          description: Response 200; application invariants are defined in the Feature005 contracts.
        '204':
          headers:
            Cache-Control:
              schema:
                const: no-store
          description: Response 204; application invariants are defined in the Feature005 contracts.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/Capacity'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
      summary: Observe the unchanged four lifecycle event envelopes
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Response BadRequest; application invariants are defined in the Feature005 contracts.
    Unauthorized:
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Response Unauthorized; application invariants are defined in the Feature005 contracts.
    NotFound:
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Response NotFound; application invariants are defined in the Feature005 contracts.
    Gone:
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Response Gone; application invariants are defined in the Feature005 contracts.
    Capacity:
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Response Capacity; application invariants are defined in the Feature005 contracts.
    Unavailable:
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
        Location:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: '503: existing service failures or CITATION_DATA_UNAVAILABLE for a malformed persisted citation
        snapshot. No false empty-citation success.'
    InternalError:
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      description: Response InternalError; application invariants are defined in the Feature005 contracts.
  schemas:
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required:
      - input
      properties:
        input:
          type: string
          minLength: 1
          maxLength: 8000
        sessionId:
          type:
          - string
          - 'null'
          minLength: 1
          maxLength: 36
        requireEvidence:
          type: boolean
          default: false
    RunAccepted:
      type: object
      additionalProperties: false
      required:
      - taskId
      - runId
      - sessionId
      - status
      properties:
        taskId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        status:
          const: QUEUED
    RunStatus:
      type: string
      enum:
      - QUEUED
      - RUNNING
      - SUCCEEDED
      - FAILED
      - CANCELLED
      - TIMED_OUT
      - BUDGET_EXCEEDED
    TerminationReason:
      type: string
      enum:
      - COMPLETED
      - INVALID_INPUT
      - INVALID_DECISION
      - MODEL_ERROR
      - UNKNOWN_TOOL
      - DISABLED_TOOL
      - INVALID_TOOL_ARGUMENTS
      - TOOL_ERROR
      - TOOL_RESULT_INVALID
      - TOOL_RESULT_TOO_LARGE
      - CANCELLED
      - TIMED_OUT
      - BUDGET_EXCEEDED
      - QUEUE_TIMEOUT
      - DISPATCH_REJECTED
      - PERSISTENCE_UNAVAILABLE
      - INTERNAL_ERROR
      - INVALID_RUNTIME_RESULT
      - OUTPUT_TOO_LARGE
      - PROCESS_INTERRUPTED
      - CONTEXT_BUDGET_EXCEEDED
      - CONTEXT_SOURCE_UNAVAILABLE
      - LEGACY_FAILURE
      - CITATION_INVALID
      - INSUFFICIENT_EVIDENCE
    Usage:
      type: object
      additionalProperties: false
      required:
      - promptTokens
      - completionTokens
      - totalTokens
      properties:
        promptTokens:
          type: integer
          minimum: 0
          maximum: 2147483647
        completionTokens:
          type: integer
          minimum: 0
          maximum: 2147483647
        totalTokens:
          type: integer
          minimum: 0
          maximum: 2147483647
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - INVALID_REQUEST
          - UNAUTHORIZED
          - NOT_FOUND
          - RATE_LIMITED
          - DEPENDENCY_UNAVAILABLE
          - RUN_CAPACITY_EXCEEDED
          - SUBSCRIPTION_CAPACITY_EXCEEDED
          - INVALID_EVENT_CURSOR
          - EVENT_HISTORY_UNAVAILABLE
          - PERSISTENCE_UNAVAILABLE
          - SERVICE_STOPPING
          - INTERNAL_ERROR
        message:
          type: string
          maxLength: 200
        taskId:
          type: string
          format: uuid
...
    RunSnapshot:
      type: object
      additionalProperties: false
      required:
      - taskId
      - runId
      - sessionId
      - status
      - input
      - finalAnswer
      - createdAt
      - updatedAt
      - startedAt
      - finishedAt
      - cancelRequested
      - terminationReason
      - runtimeReason
      - errorCode
      - recordingComplete
      - usage
      - requireEvidence
      - citations
      properties:
        taskId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/RunStatus'
        input:
          type: string
          maxLength: 8000
        finalAnswer:
          anyOf:
          - type: string
            maxLength: 65536
          - type: 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        finishedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        cancelRequested:
          type: boolean
        terminationReason:
          anyOf:
          - $ref: '#/components/schemas/TerminationReason'
          - type: 'null'
        runtimeReason:
          anyOf:
          - type: string
            enum:
            - COMPLETED
            - INVALID_INPUT
            - INVALID_DECISION
            - MODEL_ERROR
            - UNKNOWN_TOOL
            - DISABLED_TOOL
            - INVALID_TOOL_ARGUMENTS
            - TOOL_ERROR
            - TOOL_RESULT_INVALID
            - TOOL_RESULT_TOO_LARGE
            - CANCELLED
            - TIMED_OUT
            - BUDGET_EXCEEDED
            - CONTEXT_BUDGET_EXCEEDED
            - CITATION_INVALID
            - INSUFFICIENT_EVIDENCE
          - type: 'null'
        errorCode:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
        recordingComplete:
          type: boolean
        usage:
          anyOf:
          - $ref: '#/components/schemas/Usage'
          - type: 'null'
        requireEvidence:
          type: boolean
        citations:
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/Citation'
      allOf:
      - if:
          properties:
            status:
              enum:
              - QUEUED
              - RUNNING
          required:
          - status
        then:
          properties:
            finishedAt:
              type: 'null'
            terminationReason:
              type: 'null'
            runtimeReason:
              type: 'null'
            errorCode:
              type: 'null'
            finalAnswer:
              type: 'null'
            usage:
              type: 'null'
            recordingComplete:
              const: false
      - if:
          properties:
            status:
              enum:
              - QUEUED
          required:
          - status
        then:
          properties:
            startedAt:
              type: 'null'
      - if:
          properties:
            status:
              enum:
              - RUNNING
          required:
          - status
        then:
          properties:
            startedAt:
              type: string
              format: date-time
      - if:
          properties:
            status:
              enum:
              - SUCCEEDED
              - FAILED
              - CANCELLED
              - TIMED_OUT
              - BUDGET_EXCEEDED
          required:
          - status
        then:
          properties:
            finishedAt:
              type: string
              format: date-time
            terminationReason:
              type: string
      - if:
          properties:
            status:
              enum:
              - SUCCEEDED
          required:
          - status
        then:
          properties:
            finalAnswer:
              type: string
              minLength: 1
            terminationReason:
              const: COMPLETED
            errorCode:
              type: 'null'
      - if:
          properties:
            status:
              enum:
              - FAILED
              - CANCELLED
              - TIMED_OUT
              - BUDGET_EXCEEDED
          required:
          - status
        then:
          properties:
            finalAnswer:
              type: 'null'
            errorCode:
              type: string
              minLength: 1
      - if:
          properties:
            status:
              enum:
              - CANCELLED
          required:
          - status
        then:
          properties:
            terminationReason:
              const: CANCELLED
      - if:
          properties:
            status:
              enum:
              - TIMED_OUT
          required:
          - status
        then:
          properties:
            terminationReason:
              enum:
              - TIMED_OUT
              - QUEUE_TIMEOUT
      - if:
          properties:
            status:
              enum:
              - BUDGET_EXCEEDED
          required:
          - status
        then:
          properties:
            terminationReason:
              enum:
              - BUDGET_EXCEEDED
              - CONTEXT_BUDGET_EXCEEDED
      - if:
          properties:
            status:
              enum:
              - FAILED
          required:
          - status
        then:
          properties:
            terminationReason:
              enum:
              - INVALID_INPUT
              - INVALID_DECISION
              - MODEL_ERROR
              - UNKNOWN_TOOL
              - DISABLED_TOOL
              - INVALID_TOOL_ARGUMENTS
              - TOOL_ERROR
              - TOOL_RESULT_INVALID
              - TOOL_RESULT_TOO_LARGE
              - CONTEXT_SOURCE_UNAVAILABLE
              - DISPATCH_REJECTED
              - PERSISTENCE_UNAVAILABLE
              - INTERNAL_ERROR
              - INVALID_RUNTIME_RESULT
              - OUTPUT_TOO_LARGE
              - PROCESS_INTERRUPTED
              - LEGACY_FAILURE
              - CITATION_INVALID
              - INSUFFICIENT_EVIDENCE
      - if:
          properties:
            status:
              const: SUCCEEDED
        else:
          properties:
            citations:
              maxItems: 0
      - if:
          properties:
            status:
              const: SUCCEEDED
            requireEvidence:
              const: true
          required:
          - status
          - requireEvidence
        then:
          properties:
            citations:
              minItems: 1
    StepResponse:
      type: object
      additionalProperties: false
      required:
      - stepNo
      - stepType
      - toolName
      - status
      - output
      - errorMessage
      - decisionId
      - callId
      - errorCode
      - terminal
      - latencyMs
      - promptTokens
      - completionTokens
      properties:
        stepNo:
          type: integer
          minimum: 1
          maximum: 256
        stepType:
          type: string
          enum:
          - PLANNER
          - MEMORY
          - RAG
          - LLM
          - CONTEXT_ASSEMBLY
          - MODEL_DECISION
          - TOOL
          - TOOL_CALL
          - TOOL_RESULT
          - HUMAN_APPROVAL
          - FINAL
          - FAILURE
          - TERMINATION
          - CITATION_VALIDATION
        toolName:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
        status:
          type: string
          enum:
          - RUNNING
          - SUCCESS
          - FAILED
          - SKIPPED
        output:
          anyOf:
          - type: string
            maxLength: 1024
          - type: 'null'
        errorMessage:
          anyOf:
          - type: string
            maxLength: 1024
          - type: 'null'
        decisionId:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
        callId:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
        errorCode:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
        terminal:
          type: boolean
        latencyMs:
          type: integer
          minimum: 0
          format: int64
        promptTokens:
          anyOf:
          - type: integer
            minimum: 0
            maximum: 2147483647
          - type: 'null'
        completionTokens:
          anyOf:
          - type: integer
            minimum: 0
            maximum: 2147483647
          - type: 'null'
...
    Citation:
      type: object
      additionalProperties: false
      required:
      - id
      - snapshotId
      - docId
      - documentVersion
      - chunkId
      - contentHash
      - sourcePath
      - title
      - start
      - end
      - excerpt
      properties:
        id:
          type: string
          pattern: ^S(?:[1-9]|[12][0-9]|3[0-2])$
        snapshotId: &id001
          type: string
          pattern: ^[0-9a-f]{64}$
        docId: *id001
        documentVersion: *id001
        chunkId: *id001
        contentHash: *id001
        sourcePath:
          type: string
          minLength: 1
          maxLength: 256
          pattern: ^(?!/)(?!.*(?:^|/)\.{1,2}(?:/|$))(?!.*[\\:\u0000-\u001f])[^/]+(?:/[^/]+)*$
        title:
          type: string
          minLength: 1
          maxLength: 120
        start:
          type: integer
          minimum: 0
          maximum: 1048576
        end:
          type: integer
          minimum: 1
          maximum: 1048576
        excerpt:
          type: string
          minLength: 1
          maxLength: 800
      description: Source snapshot. Runtime checks code-point range, SHA-256, unique id and UTF-16 metadata lengths;
        schema alone does not prove provenance.
...